Move entities in the domain package
