--- src/main/java/com/vaadin/demo/ce/Role.java
@@ -1,8 +1,0 @@
-/*
- */
-
-package com.vaadin.demo.ce;
-
-public enum Role {
-    ADMIN, USER;
-}
