+++ src/main/java/com/vaadin/demo/domain/Role.java
@@ -1,0 +1,8 @@
+package com.vaadin.demo.domain;
+
+/**
+ * Role entity class used in documentation examples.
+ */
+public enum Role {
+    ADMIN, USER;
+}
